+++ FO-DICOM.Core/Network/Client/Advanced/Connection/AdvancedDicomClientConnection.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) 2012-2025 fo-dicom contributors.
+﻿// Copyright (c) 2012-2026 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 #nullable disable
 
@@ -63,6 +63,7 @@ namespace FellowOakDicom.Network.Client.Advanced.Connection
             Options = dicomServiceOptions ?? throw new ArgumentNullException(nameof(dicomServiceOptions));
             _eventCollector = eventCollector ?? throw new ArgumentNullException(nameof(eventCollector));
             NetworkStream = networkStream ?? throw new ArgumentNullException(nameof(networkStream));
+            RunsAsServer = false;
         }
 
         public void StartListener()