+++ FO-DICOM.Core/Network/IDicomServer.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) 2012-2025 fo-dicom contributors.
+﻿// Copyright (c) 2012-2026 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 #nullable disable
 
@@ -83,6 +83,8 @@ namespace FellowOakDicom.Network
         /// Stop server from further listening.
         /// </summary>
         void Stop();
+
+        int GetNumberOfConnectedClients();
 
         #endregion
     }