+++ Tests/FO-DICOM.Tests/Imaging/Render/OverlayGraphicTest.cs
@@ -1,4 +1,4 @@
-// Copyright (c) 2012-2025 fo-dicom contributors.
+// Copyright (c) 2012-2026 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 #nullable disable
 
@@ -10,7 +10,7 @@ namespace FellowOakDicom.Tests.Imaging.Render
     [Collection(TestCollections.ImageSharp)]
     public class OverlayGraphicTest
     {
-        [Fact]
+        [FactForNetCore]
         public void RenderOutOfBoundsOrigin()
         {
             // Arrange