+++ Tests/FO-DICOM.Tests/Network/DicomServerTest.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) 2012-2025 fo-dicom contributors.
+﻿// Copyright (c) 2012-2026 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 #nullable disable
 
@@ -605,6 +605,11 @@ namespace FellowOakDicom.Tests.Network
                 // Wait for the server to shut down gracefully
                 await server.Registration.Task;
             }
+
+            // Wait for the ContinueWith-Task to be executed
+            // In previous versions, the server.Registration.Task also included the RemoveUnusedServicesAsync.
+            // This method now no longer exists, so the server.Registration.Task only awaits the StartAsync method, which listens and accepts clients.
+            await Task.Delay(1000); //Wait a bit more than to be sure all disconnected services are cleaned up
 
             var uniqueDisposedServices = new HashSet<DicomService>(disposedDicomServices);
             Assert.Single(uniqueDisposedServices);