+++ Tests/FO-DICOM.Tests/TestAttributes.cs
@@ -1,7 +1,8 @@
-﻿// Copyright (c) 2012-2025 fo-dicom contributors.
+﻿// Copyright (c) 2012-2026 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 #nullable disable
 
+using System.Runtime.InteropServices;
 using Xunit;
 
 namespace FellowOakDicom.Tests
@@ -27,6 +28,14 @@ namespace FellowOakDicom.Tests
 #if NET462
             Skip = "Do not run in net462";
 #endif
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
+            {
+                // TODO: all codec-related tests are failing on macos because the runner cannot load the native dlls.
+                // but this is not related to fo-dicom, but to codec package.
+                // So until this is fixed, these tests are exclued.
+                Skip = "Do not run on MacOS";
+            }
+            
         }
 
     }